+++ lib/bdi/event_base.cpp
@@ -7,7 +7,7 @@
 
 #include "event_base.h"
 
-EventBase::EventBase(int size)
+EventBase::EventBase(uint8_t size)
 {
   _size = size;
   _pending_events.reserve(size);
